chore: move bump-create-app to prepare-release
